requirements.txt
Flask==3.0.0
flask-cors==4.0.0
layoutparser[layoutmodels]==0.3.4
Pillow<10.0.0
numpy==1.24.3
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@v0.4

